requirements.txt
arxiv==1.4.8
python-dotenv==1.0.0 
scholarly==1.7.11
